+++ .storybook/stories/components/index.ts
@@ -1,0 +1,1 @@
+export { default as ResponsiveComponent } from "./ResponsiveComponent";